stop using submodules for hugo themes
The Github integration with Code Pipeline does not supply a
git repository, just a zip file of code. This makes using git
submodules difficult as it would require reconstructing the
git repo at build time and then syncing the submodule.

Instread, I'll just remove the .git/ directory in the theme and
vendor the code in my repo.