add smol theme
NOTE: I git clone'd this per the Hugo instructions, but this
creates a repo within a repo. Submodules add extra complexity
when trying to build in the pipeline. The fix is just to remove the .git
directory in the theme to vendor the code in this repo.

tl;dr -> git clone new theme and then remove .git dir in theme before
checking the code in.